♻️ refactor(selector): move remaining selector code
