+++ src/MineJason.Extras/Selectors/EntitySelectorArgumentBuilder.cs
@@ -1,16 +1,17 @@
 // SPDX-FileCopyrightText: (C) WithLithum & contributors 2023-2026
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-namespace MineJason.Extras.Selectors;
-
 using System.Globalization;
 using System.Text;
+using MineJason.Data;
+
+namespace MineJason.Extras.Selectors;
 
 internal class EntitySelectorArgumentBuilder
 {
     private bool _firstWordDone;
     private readonly StringBuilder _builder = new();
-    
+
     public void Comma()
     {
         if (_firstWordDone)
@@ -33,12 +34,12 @@ internal class EntitySelectorArgumentBuilder
     {
         WritePair(key, value.ToString(CultureInfo.InvariantCulture));
     }
-    
+
     public void WritePair(string key, DistanceRange value)
     {
         WritePair(key, value.ToString());
     }
-    
+
     public void WritePair(string key, IntegralRange value)
     {
         WritePair(key, value.ToString());
