+++ src/MineJason.Extras/Selectors/EntitySelectorParser.cs
@@ -1,12 +1,12 @@
 // SPDX-FileCopyrightText: (C) WithLithum & contributors 2023-2026
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-namespace MineJason.Extras.Selectors;
-
 using System.Collections.ObjectModel;
 using System.Globalization;
-using System.Text;
+using MineJason.Data;
 using MineJason.Extras.Selectors.Matching;
+
+namespace MineJason.Extras.Selectors;
 
 /// <summary>
 /// Provides parsing service for entity target selectors.
