+++ src/MineJason.Extras/Serialization/Json/EntityGuidSelectorConverter.cs
@@ -3,6 +3,7 @@
 
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using MineJason.Extras.Selectors;
 
 namespace MineJason.Extras.Serialization.Json;
