+++ src/MineJason/MineJason.csproj
@@ -58,4 +58,8 @@
 		</Compile>
 	</ItemGroup>
 
+	<ItemGroup>
+	  <Folder Include="Exceptions\" />
+	</ItemGroup>
+
 </Project>
